backend/requirements.txt
waitress==3.0.0
flask==2.3.2
flask_cors
flask_socketio
cryptography
psycopg2==2.9.6
python-dotenv==1.0.0